+++ src/y22d16/valve.rs
@@ -1,6 +1,6 @@
 #[derive(Debug)]
 pub struct Valve {
-    pub rate: u32,
+    pub rate: u16,
     pub paths: Vec<usize>,
     pub name: usize,
 }
